Fixing compilation issues.
