+++ types.h
@@ -1,8 +1,9 @@
 #ifndef TYPES
 #define TYPES
 
-#include <string>
 #include <fstream>
+#include <random>
+#include <string>
 #include <vector>
 
 // ******************************************************************************
@@ -36,7 +37,7 @@ struct Spin
 
         Spin(float _x, float _y, float _z);
 
-        friend std::ostream& operator<< (std::ostream& os, Spin& s);
+        friend std::ostream& operator<< (std::ostream& os, const Spin& s);
         friend float operator* (const Spin& a, const Spin& b);
 
         static Spin up ();
@@ -65,9 +66,9 @@ struct SpinGenerator
 {
     int generator;
 
-    SpinGenerator (int _case) : generator(_case) {}
+    SpinGenerator (int _case);
 
-    Spin operator() ();
+    Spin operator() () const;
 };
 
 // ******************************************************************************
